BoardGame table changes, changelogs generation standarisation

+++ src/main/java/com/example/demo/BoardGameRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BoardGameRepo extends CrudRepository<BoardGame, Long> {
+public interface BoardGameRepo extends CrudRepository<BoardGame, Integer> {
 }
